requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.24.0
neo4j>=5.15.0
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.7.0
python-multipart>=0.0.18
typing-extensions>=4.10.0 